requirements.txt
spotipy
appdirs
python-dotenv
